# Don't create needless arrays.