--- a/src/gl/lineFeature.js
+++ b/src/gl/lineFeature.js
@@ -31,79 +31,79 @@ geo.gl.lineFeature = function (arg) {
       s_update = this._update;
 
   function createVertexShader() {
-      var vertexShaderSource = [
-        '#ifdef GL_ES',
-        '  precision highp float;',
-        '#endif',
-        'attribute vec3 pos;',
-        'attribute vec3 prev;',
-        'attribute vec3 next;',
-        'attribute float offset;',
+    var vertexShaderSource = [
+      '#ifdef GL_ES',
+      '  precision highp float;',
+      '#endif',
+      'attribute vec3 pos;',
+      'attribute vec3 prev;',
+      'attribute vec3 next;',
+      'attribute float offset;',
 
-        'attribute vec3 strokeColor;',
-        'attribute float strokeOpacity;',
-        'attribute float strokeWidth;',
+      'attribute vec3 strokeColor;',
+      'attribute float strokeOpacity;',
+      'attribute float strokeWidth;',
 
-        'uniform mat4 modelViewMatrix;',
-        'uniform mat4 projectionMatrix;',
-        'uniform float pixelWidth;',
+      'uniform mat4 modelViewMatrix;',
+      'uniform mat4 projectionMatrix;',
+      'uniform float pixelWidth;',
 
-        'varying vec3 strokeColorVar;',
-        'varying float strokeWidthVar;',
-        'varying float strokeOpacityVar;',
+      'varying vec3 strokeColorVar;',
+      'varying float strokeWidthVar;',
+      'varying float strokeOpacityVar;',
 
-        'void main(void)',
-        '{',
-        /* If any vertex has been deliberately set to a negative opacity,
-         * skip doing computations on it. */
-        '  if (strokeOpacity < 0.0) {',
-        '    gl_Position = vec4(2, 2, 0, 1);',
-        '    return;',
-        '  }',
-        '  const float PI = 3.14159265358979323846264;',
-        '  vec4 worldPos = projectionMatrix * modelViewMatrix * vec4(pos.xyz, 1);',
-        '  if (worldPos.w != 0.0) {',
-        '    worldPos = worldPos/worldPos.w;',
-        '  }',
-        '  vec4 worldNext = projectionMatrix * modelViewMatrix * vec4(next.xyz, 1);',
-        '  if (worldNext.w != 0.0) {',
-        '    worldNext = worldNext/worldNext.w;',
-        '  }',
-        '  vec4 worldPrev = projectionMatrix* modelViewMatrix * vec4(prev.xyz, 1);',
-        '  if (worldPrev.w != 0.0) {',
-        '    worldPrev = worldPrev/worldPrev.w;',
-        '  }',
-        '  strokeColorVar = strokeColor;',
-        '  strokeWidthVar = strokeWidth;',
-        '  strokeOpacityVar = strokeOpacity;',
-        '  vec2 deltaNext = worldNext.xy - worldPos.xy;',
-        '  vec2 deltaPrev = worldPos.xy - worldPrev.xy;',
-        '  float angleNext = PI * 0.5;',
-        '  if (deltaNext.y < 0.0) { angleNext = -angleNext; } ',
-        '  if (deltaNext.x != 0.0) {',
-        '    angleNext = atan(deltaNext.y, deltaNext.x);',
-        '  }',
-        '  float anglePrev = PI * 0.5;',
-        '  if (deltaPrev.y < 0.0) { anglePrev = -anglePrev; } ',
-        '  if (deltaPrev.x != 0.0) {',
-        '    anglePrev = atan(deltaPrev.y, deltaPrev.x);',
-        '  }',
-        '  if (deltaPrev.xy == vec2(0.0, 0.0)) anglePrev = angleNext;',
-        '  if (deltaNext.xy == vec2(0.0, 0.0)) angleNext = anglePrev;',
-        '  float angle = (anglePrev + angleNext) / 2.0;',
-        '  float cosAngle = cos(anglePrev - angle);',
-        '  if (cosAngle < 0.1) { cosAngle = sign(cosAngle) * 1.0; angle = 0.0; }',
-        '  float distance = (offset * strokeWidth * pixelWidth) /',
-        '                    cosAngle;',
-        '  worldPos.x += distance * sin(angle);',
-        '  worldPos.y -= distance * cos(angle);',
-        '  gl_Position = worldPos;',
-        '}'
-      ].join('\n'),
-      shader = new vgl.shader(gl.VERTEX_SHADER);
-      shader.setShaderSource(vertexShaderSource);
-      return shader;
-    }
+      'void main(void)',
+      '{',
+      /* If any vertex has been deliberately set to a negative opacity,
+       * skip doing computations on it. */
+      '  if (strokeOpacity < 0.0) {',
+      '    gl_Position = vec4(2, 2, 0, 1);',
+      '    return;',
+      '  }',
+      '  const float PI = 3.14159265358979323846264;',
+      '  vec4 worldPos = projectionMatrix * modelViewMatrix * vec4(pos.xyz, 1);',
+      '  if (worldPos.w != 0.0) {',
+      '    worldPos = worldPos/worldPos.w;',
+      '  }',
+      '  vec4 worldNext = projectionMatrix * modelViewMatrix * vec4(next.xyz, 1);',
+      '  if (worldNext.w != 0.0) {',
+      '    worldNext = worldNext/worldNext.w;',
+      '  }',
+      '  vec4 worldPrev = projectionMatrix* modelViewMatrix * vec4(prev.xyz, 1);',
+      '  if (worldPrev.w != 0.0) {',
+      '    worldPrev = worldPrev/worldPrev.w;',
+      '  }',
+      '  strokeColorVar = strokeColor;',
+      '  strokeWidthVar = strokeWidth;',
+      '  strokeOpacityVar = strokeOpacity;',
+      '  vec2 deltaNext = worldNext.xy - worldPos.xy;',
+      '  vec2 deltaPrev = worldPos.xy - worldPrev.xy;',
+      '  float angleNext = PI * 0.5;',
+      '  if (deltaNext.y < 0.0) { angleNext = -angleNext; } ',
+      '  if (deltaNext.x != 0.0) {',
+      '    angleNext = atan(deltaNext.y, deltaNext.x);',
+      '  }',
+      '  float anglePrev = PI * 0.5;',
+      '  if (deltaPrev.y < 0.0) { anglePrev = -anglePrev; } ',
+      '  if (deltaPrev.x != 0.0) {',
+      '    anglePrev = atan(deltaPrev.y, deltaPrev.x);',
+      '  }',
+      '  if (deltaPrev.xy == vec2(0.0, 0.0)) anglePrev = angleNext;',
+      '  if (deltaNext.xy == vec2(0.0, 0.0)) angleNext = anglePrev;',
+      '  float angle = (anglePrev + angleNext) / 2.0;',
+      '  float cosAngle = cos(anglePrev - angle);',
+      '  if (cosAngle < 0.1) { cosAngle = sign(cosAngle) * 1.0; angle = 0.0; }',
+      '  float distance = (offset * strokeWidth * pixelWidth) /',
+      '                    cosAngle;',
+      '  worldPos.x += distance * sin(angle);',
+      '  worldPos.y -= distance * cos(angle);',
+      '  gl_Position = worldPos;',
+      '}'
+    ].join('\n'),
+    shader = new vgl.shader(gl.VERTEX_SHADER);
+    shader.setShaderSource(vertexShaderSource);
+    return shader;
+  }
 
   function createFragmentShader() {
     var fragmentShaderSource = [
@@ -123,162 +123,108 @@ geo.gl.lineFeature = function (arg) {
   }
 
   function createGLLines() {
-    var i = null,
-        j = null,
-        k = null,
-        v,
-        prev = [],
-        next = [],
-        numPts = m_this.data().length,
-        itemIndex = 0,
-        lineItemIndex = 0,
-        lineItem = null,
-        currIndex = null,
-        lineSegments = [],
-        pos = null,
-        posTmp = null,
-        strkColor = null,
-        start = null,
+    var data = m_this.data(),
+        i, j, k, v,
+        numSegments = 0,
+        lineItem, lineItemData,
+        vert = [{}, {}], vertTemp,
+        pos, posIdx3,
         position = [],
-        strkWidthArr = [],
-        strkColorArr = [],
-        strkOpacityArr = [],
-        geom = vgl.geometryData(),
         posFunc = m_this.position(),
         strkWidthFunc = m_this.style.get('strokeWidth'),
         strkColorFunc = m_this.style.get('strokeColor'),
         strkOpacityFunc = m_this.style.get('strokeOpacity'),
-        buffers = vgl.DataBuffers(1024),
-        // Sources
-        posData = vgl.sourceDataP3fv({'name': 'pos'}),
-        prvPosData = vgl.sourceDataAnyfv(
-            3, vgl.vertexAttributeKeysIndexed.Four, {'name': 'prev'}),
-        nxtPosData = vgl.sourceDataAnyfv(
-            3, vgl.vertexAttributeKeysIndexed.Five, {'name': 'next'}),
-        offPosData = vgl.sourceDataAnyfv(
-            1, vgl.vertexAttributeKeysIndexed.Six, {'name': 'offset'}),
-        strkWidthData = vgl.sourceDataAnyfv(
-            1, vgl.vertexAttributeKeysIndexed.One, {'name': 'strokeWidth'}),
-        strkColorData = vgl.sourceDataAnyfv(
-            3, vgl.vertexAttributeKeysIndexed.Two, {'name': 'strokeColor'}),
-        strkOpacityData = vgl.sourceDataAnyfv(
-            1, vgl.vertexAttributeKeysIndexed.Three,
-            {'name': 'strokeOpacity'}),
-        // Primitive indices
-        triangles = vgl.triangles(),
         order = m_this.featureVertices(),
-        addVert = function (prevPos, currPos, nextPos, offset,
-                            width, color, opacity) {
-          buffers.write('prev', prevPos, currIndex, 1);
-          buffers.write('pos', currPos, currIndex, 1);
-          buffers.write('next', nextPos, currIndex, 1);
-          buffers.write('offset', [offset], currIndex, 1);
-          buffers.write('indices', [currIndex], currIndex, 1);
-          buffers.write('strokeWidth', [width], currIndex, 1);
-          buffers.write('strokeColor', color, currIndex, 1);
-          buffers.write('strokeOpacity', [opacity], currIndex, 1);
-          currIndex += 1;
-        };
+        buffers,
+        posBuf, nextBuf, prevBuf, offsetBuf, indicesBuf,
+        strokeWidthBuf, strokeColorBuf, strokeOpacityBuf,
+        dest, dest3,
+        geom = m_mapper.geometryData();
 
-    m_this.data().forEach(function (item) {
-      lineItem = m_this.line()(item, itemIndex);
-      lineSegments.push(lineItem.length);
-      lineItem.forEach(function (lineItemData) {
-        pos = posFunc(lineItemData, lineItemIndex, item, itemIndex);
+    for (i = 0; i < data.length; i += 1) {
+      lineItem = m_this.line()(data[i], i);
+      numSegments += lineItem.length - 1;
+      for (j = 0; j < lineItem.length; j += 1) {
+        pos = posFunc(lineItem[j], j, lineItem, i);
         if (pos instanceof geo.latlng) {
-          position.push([pos.x(), pos.y(), 0.0]);
+          position.push(pos.x());
+          position.push(pos.y());
+          position.push(0.0);
         } else {
-          position.push([pos.x, pos.y, pos.z || 0.0]);
+          position.push(pos.x);
+          position.push(pos.y);
+          position.push(pos.z || 0.0);
         }
-        strkWidthArr.push(strkWidthFunc(lineItemData, lineItemIndex,
-                                        item, itemIndex));
-        strkColor = strkColorFunc(lineItemData, lineItemIndex,
-                                  item, itemIndex);
-        strkColorArr.push([strkColor.r, strkColor.g, strkColor.b]);
-        strkOpacityArr.push(strkOpacityFunc(lineItemData, lineItemIndex,
-                                            item, itemIndex));
-
-        // Assuming that we will have atleast two points
-        if (lineItemIndex === 0) {
-          posTmp = position[position.length - 1];
-          prev.push(posTmp);
-        } else {
-          prev.push(position[position.length - 2]);
-          next.push(position[position.length - 1]);
-        }
-
-        lineItemIndex += 1;
-      });
-      next.push(position[position.length - 1]);
-      lineItemIndex = 0;
-      itemIndex += 1;
-    });
+      }
+    }
 
     position = geo.transform.transformCoordinates(
                  m_this.gcs(), m_this.layer().map().gcs(),
                  position, 3);
-    prev = geo.transform.transformCoordinates(
-                 m_this.gcs(), m_this.layer().map().gcs(),
-                 prev, 3);
-    next = geo.transform.transformCoordinates(
-                 m_this.gcs(), m_this.layer().map().gcs(),
-                 next, 3);
 
-    buffers.create('pos', 3);
-    buffers.create('next', 3);
-    buffers.create('prev', 3);
-    buffers.create('offset', 1);
-    buffers.create('indices', 1);
-    buffers.create('strokeWidth', 1);
-    buffers.create('strokeColor', 3);
-    buffers.create('strokeOpacity', 1);
+    buffers = vgl.DataBuffers(numSegments * order.length);
+    posBuf           = buffers.create('pos', 3);
+    nextBuf          = buffers.create('next', 3);
+    prevBuf          = buffers.create('prev', 3);
+    offsetBuf        = buffers.create('offset', 1);
+    indicesBuf       = buffers.create('indices', 1);
+    strokeWidthBuf   = buffers.create('strokeWidth', 1);
+    strokeColorBuf   = buffers.create('strokeColor', 3);
+    strokeOpacityBuf = buffers.create('strokeOpacity', 1);
 
-    numPts = position.length;
-
-    start = buffers.alloc(numPts * 6);
-    currIndex = start;
-
-    i = 0;
-    k = 0;
-    for (j = 0; j < lineSegments.length; j += 1) {
-      i += 1;
-      for (k = 0; k < lineSegments[j] - 1; k += 1) {
-        for (v = 0; v < order.length; v += 1) {
-          addVert(prev[i + order[v][0]], position[i + order[v][0]],
-                  next[i + order[v][0]], order[v][1],
-                  strkWidthArr[i + order[v][0]],
-                  strkColorArr[i + order[v][0]],
-                  strkOpacityArr[i + order[v][0]]);
+    for (i = posIdx3 = dest = dest3 = 0; i < data.length; i += 1) {
+      lineItem = m_this.line()(data[i], i);
+      for (j = 0; j < lineItem.length; j += 1, posIdx3 += 3) {
+        lineItemData = lineItem[j];
+        /* swap entries in vert so that vert[0] is the first vertex, and
+         * vert[1] will be reused for the second vertex */
+        if (j) {
+          vertTemp = vert[0];
+          vert[0] = vert[1];
+          vert[1] = vertTemp;
         }
-        i += 1;
+        vert[1].pos = posIdx3;
+        vert[1].prev = posIdx3 - (j ? 3 : 0);
+        vert[1].next = posIdx3 + (j + 1 < lineItem.length ? 3 : 0);
+        vert[1].strokeWidth = strkWidthFunc(lineItemData, j, lineItem, i);
+        vert[1].strokeColor = strkColorFunc(lineItemData, j, lineItem, i);
+        vert[1].strokeOpacity = strkOpacityFunc(lineItemData, j, lineItem, i);
+        if (!j) {
+          continue;
+        }
+        for (k = 0; k < order.length; k += 1, dest += 1, dest3 += 3) {
+          v = vert[order[k][0]];
+          posBuf[dest3]     = position[v.pos];
+          posBuf[dest3 + 1] = position[v.pos + 1];
+          posBuf[dest3 + 2] = position[v.pos + 2];
+          prevBuf[dest3]     = position[v.prev];
+          prevBuf[dest3 + 1] = position[v.prev + 1];
+          prevBuf[dest3 + 2] = position[v.prev + 2];
+          nextBuf[dest3]     = position[v.next];
+          nextBuf[dest3 + 1] = position[v.next + 1];
+          nextBuf[dest3 + 2] = position[v.next + 2];
+          offsetBuf[dest] = order[k][1];
+          indicesBuf[dest] = dest;
+          strokeWidthBuf[dest] = v.strokeWidth;
+          strokeColorBuf[dest3]     = v.strokeColor.r;
+          strokeColorBuf[dest3 + 1] = v.strokeColor.g;
+          strokeColorBuf[dest3 + 2] = v.strokeColor.b;
+          strokeOpacityBuf[dest] = v.strokeOpacity;
+        }
       }
     }
 
-    posData.pushBack(buffers.get('pos'));
-    geom.addSource(posData);
-
-    prvPosData.pushBack(buffers.get('prev'));
-    geom.addSource(prvPosData);
-
-    nxtPosData.pushBack(buffers.get('next'));
-    geom.addSource(nxtPosData);
-
-    strkWidthData.pushBack(buffers.get('strokeWidth'));
-    geom.addSource(strkWidthData);
-
-    strkColorData.pushBack(buffers.get('strokeColor'));
-    geom.addSource(strkColorData);
-
-    strkOpacityData.pushBack(buffers.get('strokeOpacity'));
-    geom.addSource(strkOpacityData);
-
-    offPosData.pushBack(buffers.get('offset'));
-    geom.addSource(offPosData);
-
-    triangles.setIndices(buffers.get('indices'));
-    geom.addPrimitive(triangles);
-
-    m_mapper.setGeometryData(geom);
+    geom.sourceByName('pos').setData(buffers.get('pos'));
+    geom.sourceByName('prev').setData(buffers.get('prev'));
+    geom.sourceByName('next').setData(buffers.get('next'));
+    geom.sourceByName('strokeWidth').setData(buffers.get('strokeWidth'));
+    geom.sourceByName('strokeColor').setData(buffers.get('strokeColor'));
+    geom.sourceByName('strokeOpacity').setData(buffers.get('strokeOpacity'));
+    geom.sourceByName('offset').setData(buffers.get('offset'));
+    geom.primitive(0).setIndices(buffers.get('indices'));
+    geom.boundsDirty(true);
+    m_mapper.modified();
+    m_mapper.boundsDirtyTimestamp().modified();
   }
 
   ////////////////////////////////////////////////////////////////////////////
@@ -289,7 +235,7 @@ geo.gl.lineFeature = function (arg) {
    */
   ////////////////////////////////////////////////////////////////////////////
   this.featureVertices = function () {
-    return [[-1, 1], [0, -1], [-1, -1], [-1, 1], [0, 1], [0, -1]];
+    return [[0, 1], [1, -1], [0, -1], [0, 1], [1, 1], [1, -1]];
   };
 
   ////////////////////////////////////////////////////////////////////////////
@@ -322,7 +268,25 @@ geo.gl.lineFeature = function (arg) {
         strkOpacityAttr = vgl.vertexAttribute('strokeOpacity'),
         // Shader uniforms
         mviUnif = new vgl.modelViewUniform('modelViewMatrix'),
-        prjUnif = new vgl.projectionUniform('projectionMatrix');
+        prjUnif = new vgl.projectionUniform('projectionMatrix'),
+        geom = vgl.geometryData(),
+        // Sources
+        posData = vgl.sourceDataP3fv({'name': 'pos'}),
+        prvPosData = vgl.sourceDataAnyfv(
+            3, vgl.vertexAttributeKeysIndexed.Four, {'name': 'prev'}),
+        nxtPosData = vgl.sourceDataAnyfv(
+            3, vgl.vertexAttributeKeysIndexed.Five, {'name': 'next'}),
+        offPosData = vgl.sourceDataAnyfv(
+            1, vgl.vertexAttributeKeysIndexed.Six, {'name': 'offset'}),
+        strkWidthData = vgl.sourceDataAnyfv(
+            1, vgl.vertexAttributeKeysIndexed.One, {'name': 'strokeWidth'}),
+        strkColorData = vgl.sourceDataAnyfv(
+            3, vgl.vertexAttributeKeysIndexed.Two, {'name': 'strokeColor'}),
+        strkOpacityData = vgl.sourceDataAnyfv(
+            1, vgl.vertexAttributeKeysIndexed.Three,
+            {'name': 'strokeOpacity'}),
+        // Primitive indices
+        triangles = vgl.triangles();
 
     m_pixelWidthUnif =  new vgl.floatUniform('pixelWidth',
                           1.0 / m_this.renderer().width());
@@ -351,6 +315,16 @@ geo.gl.lineFeature = function (arg) {
     m_actor = vgl.actor();
     m_actor.setMaterial(m_material);
     m_actor.setMapper(m_mapper);
+
+    geom.addSource(posData);
+    geom.addSource(prvPosData);
+    geom.addSource(nxtPosData);
+    geom.addSource(strkWidthData);
+    geom.addSource(strkColorData);
+    geom.addSource(strkOpacityData);
+    geom.addSource(offPosData);
+    geom.addPrimitive(triangles);
+    m_mapper.setGeometryData(geom);
   };
 
   ////////////////////////////////////////////////////////////////////////////