… src/gl/pointFeature.js
… testing/test-cases/selenium-tests/glPointsSpeed/include.js
… testing/test-cases/selenium-tests/glPointsSpeed/testGlPointsSpeed.py
… testing/test-runners/selenium_test.py.in