We create arrays using the vgl.DataBuffers.  Except for positions, which need to be transformed, load this data buffers directly.  Prior to this, the data was loaded into javascript arrays, and then that was copied to the data buffers.

Fill buffers much more efficiently by doing it directly instead of calling buffers.write and buffers.repeat

Move code from createGLPoints to _init and from createGLLines to _init.  Don't do things more than once if they can be done just once.

The point vertexShader wasn't explicitly asking for highp floats.

Added an early-exit clause to convertColor to speed up handling colors that are already in the preferred format.

Update the reference to vgl.

I will add speed test for this in geojs in another PR.

This vastly speeds up my tests in the geoapp taxi application:

Test | Loading 1,000,000 Points | Loading 1,000,000 Lines
---|---|---
Before changes | 5.7s average | 27.6s average
After changes | 0.8s average | 4.0s average