@aashish24 2015-03-07
Very nice @manthey :+1: . Will review in detail as soon as I get some time but looks like this along the lines we discussed. Thanks.