@manthey 2015-03-09
The added speed tests for selenium may need their thresholds adjusted depending on the dashboard computer(s) that run those tests.  I've tried to pick conservative values.